Update wc-talk to WP Idea Stream 2.2.0-alpha
WP Idea Stream 2.2.0 will allow sites of a network to use the plugin to manage signups for their site.
The big benefit of it for wc-talk is that you can now create a network of sites to manage call for speakers for different WordCamp and still allow speakers who are already registered to one of the site to publish a new talk on a site he is not a member of.